[FLINK-35412][Runtime] Remove `BatchCallbackRunner`
